+++ core/src/main/java/org/openedx/core/module/download/AbstractDownloader.kt
@@ -6,6 +6,7 @@ import okhttp3.OkHttpClient
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.openedx.core.config.Config
+import org.openedx.core.presentation.dialog.IAPDialogFragment.Companion.TAG
 import org.openedx.core.utils.Logger
 import retrofit2.Retrofit
 import java.io.File
@@ -38,38 +39,45 @@ abstract class AbstractDownloader : KoinComponent {
     open suspend fun download(
         url: String,
         path: String
-    ): Boolean {
+    ): DownloadResult {
         isCanceled = false
         return try {
-            val response = downloadApi.downloadFile(url).body()
-            if (response != null) {
-                val file = File(path)
-                if (file.exists()) {
-                    file.delete()
+            val responseBody = downloadApi.downloadFile(url).body() ?: return DownloadResult.ERROR
+            initializeFile(path)
+            responseBody.byteStream().use { inputStream ->
+                FileOutputStream(File(path)).use { outputStream ->
+                    writeToFile(inputStream, outputStream)
                 }
-                file.createNewFile()
-                input = response.byteStream()
-                currentDownloadingFilePath = path
-                fos = FileOutputStream(file)
-                fos.use { output ->
-                    val buffer = ByteArray(4 * 1024)
-                    var read: Int
-                    while (input!!.read(buffer).also { read = it } != -1) {
-                        output?.write(buffer, 0, read)
-                    }
-                    output?.flush()
-                }
-                true
-            } else {
-                false
             }
+            DownloadResult.SUCCESS
         } catch (e: Exception) {
-            logger.e(throwable = e)
-            false
+            e.printStackTrace()
+            if (isCanceled) DownloadResult.CANCELED else DownloadResult.ERROR
         } finally {
-            fos?.close()
-            input?.close()
+            closeResources()
         }
+    }
+
+    private fun initializeFile(path: String) {
+        val file = File(path)
+        if (file.exists()) file.delete()
+        file.createNewFile()
+        currentDownloadingFilePath = path
+    }
+
+    private fun writeToFile(inputStream: InputStream, outputStream: FileOutputStream) {
+        val buffer = ByteArray(BUFFER_SIZE)
+        var bytesRead: Int
+        while (inputStream.read(buffer).also { bytesRead = it } != -1) {
+            outputStream.write(buffer, 0, bytesRead)
+        }
+        outputStream.flush()
+    }
+
+    private fun closeResources() {
+        fos?.close()
+        input?.close()
+        currentDownloadingFilePath = null
     }
 
 
@@ -80,7 +88,7 @@ abstract class AbstractDownloader : KoinComponent {
                 fos?.close()
                 input?.close()
             } catch (e: Exception) {
-                logger.e(throwable = e)
+                e.printStackTrace()
             }
         }
         currentDownloadingFilePath?.let {
@@ -91,7 +99,11 @@ abstract class AbstractDownloader : KoinComponent {
         }
     }
 
+    enum class DownloadResult {
+        SUCCESS, CANCELED, ERROR
+    }
+
     companion object {
-        private const val TAG = "AbstractDownloader"
+        private const val BUFFER_SIZE = 4 * 1024
     }
 }
